Break logic into separate files.
